在Windows下编写和显示UI窗体，完成crash-handler的基本框架搭建
去掉gpu、ui对media的依赖（包括gn脚本和代码引用）；

+++ lcpfw/ui/accessibility/ax_tree_source_checker.h
@@ -152,7 +152,7 @@ bool AXTreeSourceChecker<AXSourceNode>::Check(AXSourceNode node,
     if (!tree_->IsValid(child)) {
       std::string msg =
           base::StringPrintf("Node %d has an invalid child (index %d): %s\n",
-                             node_id, int{i}, NodeToString(node).c_str());
+                             node_id, int{(int)i}, NodeToString(node).c_str());
       *output = msg + *output;
       return false;
     }
