+++ lcpfw/ui/views/layout/flex_layout.cc
@@ -897,7 +897,7 @@ void FlexLayout::AllocateFlexShortageAtOrder(
     DCHECK_GT(weight, 0);
     DCHECK(deficit.is_bounded());
     const SizeBound to_deduct =
-        base::ClampRound(deficit.value() * weight / float{flex_total});
+        base::ClampRound(deficit.value() * weight / float{ (float)flex_total});
     const SizeBound new_main = flex_child.preferred_size.main() - to_deduct;
 
     // If a view would shrink smaller than its current size, go with that and
@@ -1029,7 +1029,7 @@ void FlexLayout::AllocateFlexExcessAtOrder(
     SizeBound flex_amount = remaining;
     if (remaining.is_bounded()) {
       flex_amount =
-          base::ClampCeil(remaining.value() * weight / float{flex_total});
+          base::ClampCeil(remaining.value() * weight / float{(float)flex_total});
     }
     const int old_size = flex_child.current_size.main();
     const SizeBound new_main = flex_amount + old_size;
