+++ lcpfw/ui/views/layout/flex_layout_types.cc
@@ -289,7 +289,7 @@ void Span::Center(const Span& container, const Inset1D& margins) {
 
   // Case 2: room for only part of the margins.
   if (margins.size() > remaining) {
-    float scale = float{remaining} / float{margins.size()};
+    float scale = float{ (float)remaining} / float{ (float)margins.size()};
     set_start(container.start() + std::roundf(scale * margins.leading()));
     return;
   }
